removed PlayerObject script
- removed PlayerObject script
- moved temporary keybinds into Player script
- removed script component from Players in scenes
