+++ DebuggerGame/Assets/ScriptableObjects/Collection.asset
@@ -14,10 +14,4 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   savePath: /inventory.Save
   database: {fileID: 11400000, guid: c7800d2771b93ed469aebe301938bff2, type: 2}
-  Container:
-  - ID: 0
-    arthropodData: {fileID: 11400000, guid: 7446de86841fdc84dae29570e29fe6db, type: 2}
-    amount: 0
-  - ID: 1
-    arthropodData: {fileID: 11400000, guid: 24767cd8ec21f4540a421ed38a3f5f8c, type: 2}
-    amount: 0
+  Container: []
