--- DebuggerGame/Assets/Scripts/PlayerObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ea6441f2433b1cc43abd95a4e56bed37
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
